+++ CHANGELOG.md
@@ -1,3 +1,13 @@
 * Added ores to associated common tags (#c:ores, #c:ores_in_ground/*)
 * Added mining tools to #minecraft:cluster_max_harvestables (thanks Mojang)
 * Add localized Enchantment Descriptions (based on the descriptions present in the config)
+* Add Voidtouched Glass, a cheaper alternative to Glassy Void for decoration
+* Fixed lazors not being blockable with pistons
+* Fixed engoggled sorters not dropping goggles when broken
+* Cleaved block hitboxes now round up, fixing unbreakable tiny cleaves with no hitbox
+* Continuous Projector bridges now give a speedboost and highstep
+* Snares are no longer "damageable", fixing various inter-mod jank
+* Snare and shear recipe remainders are now recognized by other mods
+* Added Transferent Fungus, a block placement solution
+* Strength and Weakness potion effects now modify snare containment time
+* Updated Lib39